requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pydantic
pydantic-settings
anyio
starlette
click
h11
greenlet
typing-extensions
typing-inspection
annotated-types
annotated-doc
email-validator
dnspython
pydantic-settings
python-jose
passlib[bcrypt]
email-validator
pydantic[email]
requests
tenacity
redis
apscheduler
python-multipart
transformers
pillow
opencv-python-headless
pytesseract
numpy
